Created ReusableComponents class, changes made according to PR Review
